Refactored validators to standardize naming conventions and move to a validator package.
